+++ changelog.md
@@ -5,6 +5,7 @@
 * Pin UID version in appveyor
 * Log errors from faf-uid executable (#741, #742)
 * Add menu option: Show paths for maps, mods & replays (#735)
+* Slightly clean up client resize logic (#747, #748)
 
 Contributors:
   - Grothe